+++ src/compiler/sema_types.c
@@ -9,7 +9,6 @@ static inline bool sema_resolve_ptr_type(SemaContext *context, TypeInfo *type_in
 static inline bool sema_resolve_array_type(SemaContext *context, TypeInfo *type, ResolveTypeKind resolve_kind);
 static inline bool sema_resolve_type(SemaContext *context, TypeInfo *type_info, ResolveTypeKind resolve_kind);
 static bool sema_resolve_type_identifier(SemaContext *context, TypeInfo *type_info, ResolveTypeKind resolve_type_kind);
-INLINE bool sema_resolve_vatype(SemaContext *context, TypeInfo *type_info);
 INLINE bool sema_resolve_typefrom(SemaContext *context, TypeInfo *type_info, ResolveTypeKind resolve_kind);
 INLINE bool sema_resolve_typeof(SemaContext *context, TypeInfo *type_info);
 static inline bool sema_check_ptr_type(SemaContext *context, TypeInfo *type_info, Type *inner);
@@ -404,19 +403,6 @@ INLINE bool sema_resolve_typefrom(SemaContext *context, TypeInfo *type_info, Res
 	UNREACHABLE
 }
 
-// $vatype(...)
-INLINE bool sema_resolve_vatype(SemaContext *context, TypeInfo *type_info)
-{
-	if (!context->macro_has_vaargs)
-	{
-		RETURN_SEMA_ERROR(type_info, "'%s' can only be used inside of a macro with untyped vaargs.", token_type_to_string(TOKEN_CT_VATYPE));
-	}
-	ASSIGN_EXPR_OR_RET(Expr *arg_expr, sema_expr_analyse_ct_arg_index(context, type_info->unresolved_type_expr, NULL), false);
-	if (!sema_analyse_expr(context, arg_expr)) return false;
-	if (arg_expr->expr_kind != EXPR_TYPEINFO) RETURN_SEMA_ERROR(arg_expr, "The argument was not a type.");
-	type_info->type = arg_expr->type_expr->type;
-	return true;
-}
 
 bool sema_unresolved_type_is_generic(SemaContext *context, TypeInfo *type_info)
 {
@@ -530,9 +516,6 @@ static inline bool sema_resolve_type(SemaContext *context, TypeInfo *type_info, 
 			UNREACHABLE
 		case TYPE_INFO_GENERIC:
 			if (!sema_resolve_generic_type(context, type_info)) return type_info_poison(type_info);
-			goto APPEND_QUALIFIERS;
-		case TYPE_INFO_VATYPE:
-			if (!sema_resolve_vatype(context, type_info)) return type_info_poison(type_info);
 			goto APPEND_QUALIFIERS;
 		case TYPE_INFO_CT_IDENTIFIER:
 		case TYPE_INFO_IDENTIFIER:
@@ -688,7 +671,7 @@ static uint32_t hash_function(Signature *sig)
 	return (uint32_t)((hash >> 16) ^ hash);
 }
 
-static inline Type *func_create_new_func_proto(Signature *sig, CallABI abi, uint32_t hash, FuncTypeEntry *entry)
+static inline Type *func_create_new_func_proto(Signature *sig, CallABI abi UNUSED, uint32_t hash, FuncTypeEntry *entry)
 {
 	unsigned param_count = vec_size(sig->params);
 	FunctionPrototype *proto = CALLOCS(FunctionPrototype);