+++ src/compiler/tokens.c
@@ -367,14 +367,8 @@ const char *token_type_to_string(TokenType type)
 			return "$include";
 		case TOKEN_CT_VACOUNT:
 			return "$vacount";
-		case TOKEN_CT_VATYPE:
-			return "$vatype";
-		case TOKEN_CT_VACONST:
-			return "$vaconst";
 		case TOKEN_CT_VAARG:
 			return "$vaarg";
-		case TOKEN_CT_VAEXPR:
-			return "$vaexpr";
 		case TOKEN_CT_VASPLAT:
 			return "$vasplat";
 		case TOKEN_CT_REFLECT:
@@ -399,6 +393,6 @@ const char *token_type_to_string(TokenType type)
 
 bool token_is_any_type(TokenType type)
 {
-	return (type >= TOKEN_VOID && type <= TOKEN_TYPEID) || type == TOKEN_CT_TYPE_IDENT || type == TOKEN_TYPE_IDENT || type == TOKEN_CT_VATYPE;
+	return (type >= TOKEN_VOID && type <= TOKEN_TYPEID) || type == TOKEN_CT_TYPE_IDENT || type == TOKEN_TYPE_IDENT;
 }