+++ releasenotes.md
@@ -51,7 +51,8 @@
 - Add `$expand` compile time function to convert strings to code.
 - Constdef now infers through unary negations.
 - Only used libraries are scanned for dependencies. #3144
-
+- `$vaconst`, `$vaexpr` and `$vatype` removed.
+ 
 ### Stdlib changes
 - `std::collections::RingBuffer` has been renamed `RingList`.
 - Add `List.remove_unordered_at`.